+++ src/server/game/Entities/GameObject/GameObject.h
@@ -612,7 +612,7 @@ struct GameObjectData
     float rotation2;
     float rotation3;
     int32  spawntimesecs;
-	int64  owner_id;
+	int32  owner_id;
     uint32 animprogress;
     GOState go_state;
     uint8 spawnMask;
@@ -654,7 +654,7 @@ class TC_GAME_API GameObject : public WorldObject, public GridObject<GameObject>
         void RemoveFromWorld() override;
         void CleanupsBeforeDelete(bool finalCleanup = true) override;
 
-        bool Create(ObjectGuid::LowType guidlow, uint32 name_id, Map* map, uint32 phaseMask, float x, float y, float z, float ang, float rotation0, float rotation1, float rotation2, float rotation3, uint32 animprogress, GOState go_state, uint32 artKit = 0, uint64 owner_id = 0);
+        bool Create(ObjectGuid::LowType guidlow, uint32 name_id, Map* map, uint32 phaseMask, float x, float y, float z, float ang, float rotation0, float rotation1, float rotation2, float rotation3, uint32 animprogress, GOState go_state, uint32 artKit = 0, uint32 owner_id = 0);
         void Update(uint32 p_time) override;
         GameObjectTemplate const* GetGOInfo() const { return m_goInfo; }
         GameObjectData const* GetGOData() const { return m_goData; }
@@ -884,7 +884,7 @@ class TC_GAME_API GameObject : public WorldObject, public GridObject<GameObject>
         GameObjectData const* m_goData;
         GameObjectValue m_goValue;
 
-		uint64 m_ownerId;
+		uint32 m_ownerId;
 
 		uint32 m_creatureAttach;
